Refactor Proposal Field Methods
This commit pulls three functions explicitly related to proposal field values
into their own utility file, for reuse in the subsequent Organization
Proposal side-panel Component, and future components that render proposal
values.

Issue #613 Add Proposal Side Panel
